refactor: remove LottoTicketsFactory and move createAutomatically method to LottoTickets
